Move Resource to Resources::Base
