+++ lib/cm_api/resources.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module CMAPI
+  module Resources
+    autoload :Base, "cm_api/resources/base"
+  end
+end
